+++ Tests/Unit/Extractor/SeoKeywordsExtractorTest.php
@@ -29,8 +29,6 @@ class SeoKeywordsExtractorTest extends BaseTestCase
     {
         return array(
             array($this->getMock('Symfony\Cmf\Bundle\SeoBundle\Extractor\SeoKeywordsReadInterface')),
-            array($this->getMock('Symfony\Cmf\Bundle\SeoBundle\Extractor\SeoDescriptionReadInterface'), false),
-            array($this->getMock('Symfony\Cmf\Bundle\SeoBundle\Extractor\SeoOriginalRouteReadInterface'), false),
             array($this->getMock('Symfony\Cmf\Bundle\SeoBundle\Model\SeoAwareInterface'), false),
         );
     }