+++ Tests/Unit/Extractor/SeoTitleExtractorTest.php
@@ -29,8 +29,6 @@ class SeoTitleExtractorTest extends BaseTestCase
     {
         return array(
             array($this->getMock('Symfony\Cmf\Bundle\SeoBundle\Extractor\SeoTitleReadInterface')),
-            array($this->getMock('Symfony\Cmf\Bundle\SeoBundle\Extractor\SeoDescriptionReadInterface'), false),
-            array($this->getMock('Symfony\Cmf\Bundle\SeoBundle\Extractor\SeoOriginalRouteReadInterface'), false),
             array($this->getMock('Symfony\Cmf\Bundle\SeoBundle\Model\SeoAwareInterface'), false),
         );
     }