+++ Tests/WebTest/Admin/SeoContentAdminExtensionTest.php
@@ -63,4 +63,15 @@ class SeoContentAdminExtensionTest extends BaseTestCase
         $this->assertCount(1, $crawler->filter('html:contains("description")'));
         $this->assertCount(1, $crawler->filter('html:contains("keywords")'));
     }
+
+    public function testExtraPropertyEditView()
+    {
+        $crawler = $this->client->request('GET', '/admin/cmf/seo/seoawarecontent/test/content/content-arbitrary-property/edit');
+        $res = $this->client->getResponse();
+
+        $this->assertEquals(200, $res->getStatusCode());
+        $this->assertCount(1, $crawler->filter('html:contains("Key")'));
+        $this->assertCount(1, $crawler->filter('html:contains("Value")'));
+        $this->assertCount(1, $crawler->filter('html:contains("Type")'));
+    }
 }